requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
filelock==3.18.0
fsspec==2025.3.2
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.1
opencv-python==4.11.0.86
packaging==24.2
Pillow==11.2.1
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
pytesseract==0.3.13
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
urllib3==2.4.0
ultralytics==8.3.111
ultralytics-thop==2.0.14
face_recognition==1.3.0
python-dotenv==1.0.0
